--- src/main/java/com/upao/eduacces/EduAccesApplication.java
@@ -1,13 +1,0 @@
-package com.upao.eduacces;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EduAccesApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EduAccesApplication.class, args);
-    }
-
-}
