--- src/main/java/com/upao/eduacces/domain/EstudianteCursoPK.java
@@ -1,22 +1,0 @@
-package com.upao.eduacces.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-import java.io.Serializable;
-
-@Data
-@Embeddable
-@EqualsAndHashCode
-public class EstudianteCursoPK implements Serializable {
-
-    @Column(name = "id_estudiante")
-    private Long idEstudiante;
-
-    @Column(name = "id_curso")
-    private Long idCurso;
-}
-
-
