--- src/main/java/com/upao/eduacces/domain/Material.java
@@ -1,33 +1,0 @@
-package com.upao.eduacces.domain;
-
-import com.upao.eduacces.enums.TipoMaterial;
-import jakarta.persistence.*;
-import lombok.Data;
-
-import java.sql.Date;
-
-@Data
-@Entity
-@Table(name = "material")
-public class Material {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "titulo", length = 100, nullable = false)
-    private String titulo;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "tipo", length = 50, nullable = false)
-    private TipoMaterial tipo;
-
-    @Column(name = "fecha_subida", nullable = false)
-    private Date fechaSubida;
-
-    @ManyToOne
-    @JoinColumn(name = "curso_id", referencedColumnName = "id", nullable = false)
-    private Curso curso;
-}
-
