--- src/main/java/com/upao/eduacces/enums/EstadoCurso.java
@@ -1,6 +1,0 @@
-package com.upao.eduacces.enums;
-
-public enum EstadoCurso {
-  PUBLICADO,
-  NO_PUBLICADO
- }
