--- src/main/java/com/upao/eduacces/enums/EstadoProgreso.java
@@ -1,8 +1,0 @@
-package com.upao.eduacces.enums;
-
-public enum EstadoProgreso {
-    INICIADO,
-    EN_PROGRESO,
-    COMPLETADO,
-    PENDIENTE
-}
