--- src/main/java/com/upao/eduacces/enums/TipoMaterial.java
@@ -1,10 +1,0 @@
-package com.upao.eduacces.enums;
-
-public enum TipoMaterial {
-    VIDEO,
-    PDF,
-    TEXTO,
-    AUDIO,
-    IMAGEN,
-    CUESTIONARIO
-}
