--- src/main/java/com/upao/eduacces/enums/TipoPago.java
@@ -1,6 +1,0 @@
-package com.upao.eduacces.enums;
-
-public enum TipoPago {
-    TARJETA_CREDITO,
-    PAYPAL,
-}
