+++ src/main/java/com/upao/eduaccess/EduAccesApplication.java
@@ -1,0 +1,13 @@
+package com.upao.eduaccess;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EduAccesApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(EduAccesApplication.class, args);
+    }
+
+}
