+++ src/main/java/com/upao/eduaccess/enums/EstadoCurso.java
@@ -1,0 +1,6 @@
+package com.upao.eduaccess.enums;
+
+public enum EstadoCurso {
+  PUBLICADO,
+  NO_PUBLICADO
+ }
