+++ src/main/java/com/upao/eduaccess/enums/EstadoProgreso.java
@@ -1,0 +1,8 @@
+package com.upao.eduaccess.enums;
+
+public enum EstadoProgreso {
+    INICIADO,
+    EN_PROGRESO,
+    COMPLETADO,
+    PENDIENTE
+}
