+++ src/main/java/com/upao/eduaccess/enums/TipoMaterial.java
@@ -1,0 +1,10 @@
+package com.upao.eduaccess.enums;
+
+public enum TipoMaterial {
+    VIDEO,
+    PDF,
+    TEXTO,
+    AUDIO,
+    IMAGEN,
+    CUESTIONARIO
+}
