+++ src/main/java/com/upao/eduaccess/enums/TipoPago.java
@@ -1,0 +1,6 @@
+package com.upao.eduaccess.enums;
+
+public enum TipoPago {
+    TARJETA_CREDITO,
+    PAYPAL,
+}
