+++ src/main/java/com/upao/eduaccess/repository/ComentarioRepository.java
@@ -1,0 +1,13 @@
+package com.upao.eduaccess.repository;
+
+import com.upao.eduaccess.domain.Comentario;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface ComentarioRepository extends JpaRepository<Comentario, Long> {
+    List<Comentario> findByCursoId(Long cursoId);
+}
+
