+++ src/main/java/com/upao/eduaccess/repository/CursoRepository.java
@@ -1,0 +1,9 @@
+package com.upao.eduaccess.repository;
+
+import com.upao.eduaccess.domain.Curso;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CursoRepository extends JpaRepository<Curso, Long> {
+}
