--- src/test/java/com/upao/eduacces/EduAccesApplicationTests.java
@@ -1,13 +1,0 @@
-package com.upao.eduacces;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EduAccesApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
