+++ src/test/java/com/upao/eduaccess/EduAccesApplicationTests.java
@@ -1,0 +1,13 @@
+package com.upao.eduaccess;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EduAccesApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
